backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
pydantic==2.6.4
pydantic-settings==2.2.1
aiofiles==23.2.1
httpx==0.27.0
reportlab==4.2.0
python-dateutil==2.9.0
